Drop support for Scala 2.12
