+++ core/src/main/scala/shapeless/syntax/singletons.scala
@@ -47,3 +47,8 @@ object SingletonOps {
       val witness = w
     }
 }
+
+object singleton {
+  implicit def mkSingletonOps[T](t: T): SingletonOps.Aux[t.type] =
+    SingletonOps.instance[t.type](Witness(t))
+}
