--- core/src/main/scala_2.13+/shapeless/syntax/singleton.scala
@@ -1,8 +1,0 @@
-package shapeless.syntax
-
-import shapeless.Witness
-
-object singleton {
-  implicit def mkSingletonOps[T](t: T): SingletonOps.Aux[t.type] =
-    SingletonOps.instance[t.type](Witness(t))
-}
