--- core/src/main/scala_2.13-/shapeless/syntax/singleton.scala
@@ -1,10 +1,0 @@
-package shapeless.syntax
-
-import shapeless.SingletonTypeMacros
-
-import scala.language.experimental.macros
-
-object singleton {
-  implicit def mkSingletonOps(t: Any): SingletonOps =
-    macro SingletonTypeMacros.mkSingletonOps
-}
